+++ module/data/_module.mjs
@@ -10,3 +10,4 @@ export {default as MistEngineFeature} from "./item-feature.mjs";
 export {default as MistEngineSpell} from "./item-spell.mjs";
 export {default as MistEngineItemThemeBook} from "./item-themebook.mjs";
 export {default as MistEngineItemBackpack} from "./item-backpack.mjs";
+export {default as MistEngineSceneData} from "./item-scene-data.mjs";