+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
-## 1.2.0
+## 0.2
 
-- Add support for Foundry v10
+- scene window dialog for handing scene and story markers / tags (found in the left sidebar under Journal Notes > Scene Tags)
+- temporary statuses for characters and npcs